final_project-main/DRF_Auth/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-environ==0.9.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
agora-token-builder==1.0.0
asgiref==3.4.1
